fix: allow anvils to set print custom names
Fixes #10. Instead of implementing Nameable and having to deal with
the consequences of two different NBT keys for custom names, we just
cause ItemStack to defer to the PrintData.label for custom names.

In fixing this, the discussion of "should print names be italic" came
up. To be consistent with the vanilla behavior of having user-generated
content be italic, all print names are now italic (hasCustomName is
always true).