feat: .2dja support for poster.lua
Add support for .2dja files, a file format used by the community to simplify the printing of large signs, which essentially contains multiple .2dj files.

Additionally, the status text has been updated.

While maintaining compatibility, two optional numbers can now be appended to the syntax to indicate the interval of pages to be used if a .2dja file is provided. Or a single specific page.

This commit also ensures that the program is now compatible with normal PCs that don't support all colors.

Finally, comments and style changes have been included for improved readability.